+++ python/ray/train/v2/tests/test_telemetry.py
@@ -26,8 +26,8 @@ def test_not_used_on_import(reset_usage_lib, callsite: TelemetryCallsite):
 
 
 @pytest.mark.parametrize("callsite", list(TelemetryCallsite))
-def test_used_on_train__fit(reset_usage_lib, callsite: TelemetryCallsite):
-    def _call_train_fit():
+def test_used_on_trainer_fit(reset_usage_lib, callsite: TelemetryCallsite):
+    def _call_trainer_fit():
         def train_fn():
             pass
 
@@ -35,13 +35,41 @@ def test_used_on_train__fit(reset_usage_lib, callsite: TelemetryCallsite):
         trainer.fit()
 
     check_library_usage_telemetry(
-        _call_train_fit,
+        _call_trainer_fit,
         callsite=callsite,
         expected_library_usages=[{"train"}, {"core", "train"}],
         expected_extra_usage_tags={
             "train_trainer": "DataParallelTrainer",
         },
     )
+
+
+@pytest.mark.skipif(
+    sys.version_info.major == 3 and sys.version_info.minor >= 12,
+    reason="Python 3.12+ does not have Tensorflow installed on CI due to dependency conflicts.",
+)
+def test_tag_train_entrypoint(mock_record):
+    """Test that Train v2 entrypoints are recorded correctly."""
+    from ray.train.v2.lightgbm.lightgbm_trainer import LightGBMTrainer
+    from ray.train.v2.tensorflow.tensorflow_trainer import TensorflowTrainer
+    from ray.train.v2.torch.torch_trainer import TorchTrainer
+    from ray.train.v2.xgboost.xgboost_trainer import XGBoostTrainer
+
+    trainer_classes = [
+        TorchTrainer,
+        TensorflowTrainer,
+        XGBoostTrainer,
+        LightGBMTrainer,
+    ]
+    for trainer_cls in trainer_classes:
+        trainer = trainer_cls(
+            lambda: None,
+            scaling_config=ray.train.ScalingConfig(num_workers=2),
+        )
+        assert (
+            mock_record[ray_usage_lib.TagKey.TRAIN_TRAINER]
+            == trainer.__class__.__name__
+        )
 
 
 if __name__ == "__main__":
